# docs/swagger.yaml
basePath: /v1
definitions:
  entity.GenericJob:
    properties:
      jobId:
        example: "12345"
        type: string
      jobName:
        example: name
        type: string
      jobStatus:
        example: jobStatus
        type: string
    type: object
  v1.createInferenceJobRequest:
    type: object
  v1.createInferenceJobResponse:
    properties:
      entryPoint:
        example: http://localhost:8080
        type: string
      jobId:
        example: "12345"
        type: string
    type: object
  v1.createTrainingJobRequest:
    properties:
      dockerImageName:
        example: yjack0000cs12/llm-training:latest
        type: string
      twccJobId:
        example: "237139"
        type: string
    type: object
  v1.createTrainingJobResponse:
    properties:
      jobId:
        example: "12345"
        type: string
    type: object
  v1.eResponse:
    properties:
      error:
        example: message
        type: string
    type: object
  v1.getInferenceJobResponse:
    properties:
      job:
        $ref: '#/definitions/entity.GenericJob'
    type: object
  v1.getJobResponse:
    properties:
      job:
        $ref: '#/definitions/entity.GenericJob'
    type: object
  v1.listInferenceJobResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/entity.GenericJob'
        type: array
    type: object
  v1.sResponse:
    properties:
      success:
        example: message
        type: string
    type: object
info:
  contact: {}
  description: swagger test example
  title: swagger test
  version: "1.0"
paths:
  /inference-jobs:
    get:
      consumes:
      - application/json
      description: list all inference jobs
      operationId: list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.listInferenceJobResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: list all inference jobs
      tags:
      - inference-jobs
    post:
      consumes:
      - application/json
      description: create inference job
      operationId: create
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/v1.createInferenceJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.createInferenceJobResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: create inference job
      tags:
      - inference-jobs
  /inference-jobs/{id}:
    delete:
      consumes:
      - application/json
      description: delete inference job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.sResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: delete inference job
      tags:
      - inference-jobs
    get:
      consumes:
      - application/json
      description: get inference job
      operationId: get
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getInferenceJobResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: get inference job
      tags:
      - inference-jobs
  /training-jobs/{id}:
    get:
      consumes:
      - application/json
      description: get training job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getJobResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: get training job
      tags:
      - training-jobs
  /training-jobs/all:
    get:
      consumes:
      - application/json
      description: list all training jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getJobResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: list all training jobs
      tags:
      - training-jobs
  /training-jobs/create:
    post:
      consumes:
      - application/json
      description: create training job
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/v1.createTrainingJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.createTrainingJobResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.eResponse'
      summary: create training job
      tags:
      - training-jobs
schemes:
- http
- https
swagger: "2.0"